Tests/LibWeb/Text/input/DOM/getElementsByClassName-empty-string.html

<!DOCTYPE html>
<script src="../include.js"></script>
<div class=""></div>
<script>
    test(() => {
        println(`document.getElementsByClassName("").length: ${document.getElementsByClassName("").length}`);
    });
</script>
