Tests/LibWeb/Text/expected/DOM/getElementsByClassName-empty-string.txt

document.getElementsByClassName("").length: 0
